+++ src/main/java/com/sandymandy/pleasurecraft/client/models/BiaModel.java
@@ -1,5 +1,0 @@
-package com.sandymandy.pleasurecraft.client.models;
-
-import com.sandymandy.pleasurecraft.entity.girls.BiaEntity;
-
-public class BiaModel extends AbstractGirlModel<BiaEntity> {}