+++ src/main/java/com/sandymandy/pleasurecraft/client/models/MikaModel.java
@@ -1,0 +1,5 @@
+package com.sandymandy.pleasurecraft.client.models;
+
+import com.sandymandy.pleasurecraft.entity.girls.MikaEntity;
+
+public class MikaModel extends AbstractGirlModel<MikaEntity> {}