+++ src/main/java/com/sandymandy/pleasurecraft/client/models/ZhongeziModel.java
@@ -1,6 +1,0 @@
-package com.sandymandy.pleasurecraft.client.models;
-
-import com.sandymandy.pleasurecraft.entity.girls.ZhongeziEntity;
-
-public class ZhongeziModel extends AbstractGirlModel<ZhongeziEntity>{
-}