+++ src/main/java/com/sandymandy/pleasurecraft/client/models/ZongModel.java
@@ -1,0 +1,6 @@
+package com.sandymandy.pleasurecraft.client.models;
+
+import com.sandymandy.pleasurecraft.entity.girls.ZongEntity;
+
+public class ZongModel extends AbstractGirlModel<ZongEntity>{
+}