+++ src/main/java/com/sandymandy/pleasurecraft/entity/girls/MikaEntity.java
@@ -13,25 +13,25 @@ import net.minecraft.world.World;
 
 import java.util.List;
 
-public class ZhongeziEntity extends SceneEntity {
+public class MikaEntity extends SceneEntity {
 
-    public ZhongeziEntity(EntityType<? extends AbstractGirlEntity> entityType, World world) {
+    public MikaEntity(EntityType<? extends AbstractGirlEntity> entityType, World world) {
         super(entityType, world);
     }
 
     @Override
     protected Item getTameItem() {
-        return Items.POPPY;
+        return Items.OPEN_EYEBLOSSOM;
     }
 
     @Override
     protected String getGirlDisplayName() {
-        return "Zhongezi";
+        return "Mika";
     }
 
     @Override
     public String getGirlID() {
-            return "zhongezi";
+        return "mika";
     }
 
     @Override
@@ -39,27 +39,7 @@ public class ZhongeziEntity extends SceneEntity {
 
     @Override
     public List<SceneOptions> getSceneOptions() {
-        return List.of(
-                SceneOptions.of("Anal",
-                        6,
-                        List.of("anal_intro"),
-                        List.of("anal_slow"),
-                        List.of("anal_fast"),
-                        "anal_cum",
-                        true,
-                        0f,
-                        List.of("anal_lay_on_bed", "anal_bed_idle")),
-
-                SceneOptions.of("Doggy",
-                        8,
-                        List.of("prone_doggy_intro"),
-                        List.of("prone_doggy_slow"),
-                        List.of("prone_doggy_hard1"/*,"prone_doggy_hard2","prone_doggy_hard3"*/),
-                        "prone_doggy_cum",
-                        true,
-                        1f,
-                        List.of("sit_down", "sit_down_idle"))
-        );
+        return List.of();
     }
 
     public static DefaultAttributeContainer.Builder createAttributes() {
@@ -68,5 +48,6 @@ public class ZhongeziEntity extends SceneEntity {
                 .add(EntityAttributes.MOVEMENT_SPEED, .20)
                 .add(EntityAttributes.TEMPT_RANGE, 5)
                 .add(EntityAttributes.ATTACK_DAMAGE, 2);
+
     }
 }