+++ src/main/java/com/sandymandy/pleasurecraft/item/PleasureCraftItemGroups.java
@@ -1,6 +1,7 @@
 package com.sandymandy.pleasurecraft.item;
 
 import com.sandymandy.pleasurecraft.PleasureCraft;
+import com.sandymandy.pleasurecraft.entity.PleasureCraftEntities;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
@@ -12,17 +13,10 @@ import net.minecraft.util.Identifier;
 public class PleasureCraftItemGroups {
     public static final ItemGroup PLEASURE_CRAFT_ITEM_GROUP = Registry.register(Registries.ITEM_GROUP,
             Identifier.of(PleasureCraft.MOD_ID, "pleasure_craft_items"),
-            FabricItemGroup.builder().icon(() -> new ItemStack(PleasureCraftItems.LUCY_SPAWN_EGG))
+            FabricItemGroup.builder().icon(() -> new ItemStack(PleasureCraftEntities.getFirstSpawnEgg()))
                     .displayName(Text.translatable("itemgroup.pleasurecraft.pleasure_craft_items"))
                     .entries((displayContext, entries) -> {
-                      entries.add(PleasureCraftItems.LUCY_SPAWN_EGG);
-                      entries.add(PleasureCraftItems.BIA_SPAWN_EGG);
-                      entries.add(PleasureCraftItems.ZHONGEZI_SPAWN_EGG);
-
-
-
-
-
+                        PleasureCraftEntities.getAllSpawnEggs().forEach(entries::add);
                     }).build());
 
     /*public static final ItemGroup PLEASURE_CRAFT_ITEM_GROUP = Registry.register(Registries.ITEM_GROUP,